Mo'Trees
Changed the way tree class works, it is a part handler rather than the
part it self. It creates a random (limited by a specified number) amount
of a chosen(from a specified array of game parts) part at random
position from the object and a random rotation.

The rotation is chosen before the instaintian using random limited
nubers using

Quaternion.Euler (0, 0, 10);